added LCN updater
