+++ index.html
@@ -36,7 +36,7 @@
     <h2>Modifying Existing Content</h2>
 		<a href="./url/src/url.html">URL Tool</a></br>
 		<a href="./recode/src/recode.html">List Recode</a></br>
-		<a href="./lcn/src/lcn.html">LCN Updater</a></br>
+		<a href="./lcn_updater/src/lcn_updater.html">LCN Updater</a></br>
 		
     <h2>Deleting Content</h2>
 		<a href="./del/src/del.html">Delete an Item/Paragraph </a></br>
